refactor: 重构web端kafka使用方式
