+++ flink-2-hbase/src/main/resources/config.properties
@@ -17,5 +17,5 @@ redis.db=0
 
 
 ## kafka config
-kafka.bootstrap.servers=192.168.1.13:9092
-kafka.zookeeper.connect=192.168.1.13:2181
+kafka.bootstrap.servers=192.168.1.7:9092
+kafka.zookeeper.connect=192.168.1.7:2181
