--- web/src/main/resources/config.properties
@@ -1,23 +1,0 @@
-## hbase config
-hbase.rootdir=hdfs://localost:9000/hbase
-hbase.zookeeper.quorum=localhost
-hbase.client.scanner.timeout.period=1000
-hbase.rpc.timeout=1000
-
-
-## mysql config
-mysql.url=jdbc:mysql://localhost:3306/con?serverTimezone=GMT%2B8
-mysql.name=root
-mysql.pass=123456
-
-## redis config 192.168.0.100
-redis.host=localhost
-redis.port=6379
-redis.db=0
-
-
-## kafka config
-kafka.bootstrap.servers=192.168.1.13:9092
-kafka.zookeeper.connect=192.168.1.13:2181
-kafka.key.serializer=org.apache.kafka.common.serialization.StringSerializer
-kafka.value.serializer=org.apache.kafka.common.serialization.StringSerializer
